+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.3] - 2020-07-04
+### Added
+- `strokeWidth`, `strokeColor`, `shadowColor`, `shadowBlur`, `shadowX` and `shadowY` properties
+- `strokeWidth`, `shadowBlur`, `shadowX` and `shadowY` are animatable
+- `color`, `strokeColor` and `shadowColor` allow any color value (hex, hsl, name, etc).
+
 ## [0.1.2] - 2020-06-27
 ### Fixed
 - `play()` does not work sometimes
@@ -19,5 +25,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2020-06-27
 First version with basic features
 
+[0.1.3]: https://github.com/oom-components/particles/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/oom-components/particles/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/oom-components/particles/compare/v0.1.0...v0.1.1